+++ accounts/urls.py
@@ -11,4 +11,5 @@ urlpatterns = [
     path("profile/update_email/", views.update_email, name='update_email'),
     path("profile/delete_account/", views.delete_account, name='delete_account'),
     path("change-password/", auth_views.PasswordChangeView.as_view(), name='change_password'),
+    path("profile/leaderboard/", views.delete_account, name='leaderboard_personal'),
 ]